requirements.txt
Django==3.2.6
selenium
chromedriver-binary
chromedriver-binary-auto
djangorestframework
